+++ esm/index.js
@@ -1,11 +1,11 @@
 /* eslint-disable node/no-unpublished-import, node/no-unsupported-features/es-syntax, import/extensions */
-
 import CRC32 from '../src/data-hash-crc32.js';
 import DataBitstream from '../src/data-bitstream.js';
 import DataBuffer from '../src/data-buffer.js';
 import DataBufferList from '../src/data-buffer-list.js';
 import DataStream from '../src/data-stream.js';
 import LZW from '../src/data-compression-lzw.js';
+import { formatBytes } from '../src/data-formating.js';
 
 export default {
   CRC32,
@@ -14,6 +14,7 @@ export default {
   DataBufferList,
   DataStream,
   LZW,
+  formatBytes,
 };
 
 export { default as CRC32 } from '../src/data-hash-crc32.js';
@@ -22,3 +23,4 @@ export { default as DataBuffer } from '../src/data-buffer.js';
 export { default as DataBufferList } from '../src/data-buffer-list.js';
 export { default as DataStream } from '../src/data-stream.js';
 export { default as LZW } from '../src/data-compression-lzw.js';
+export { formatBytes } from '../src/data-formating.js';