+++ src/data-bitstream.js
@@ -1,3 +1,4 @@
+/** @type {Function} */
 let debug = () => {}; /* istanbul ignore next */ if (process.env.UTTORI_DATA_DEBUG) { try { debug = require('debug')('DataBitstream'); } catch {} }
 const DataStream = require('./data-stream');
 const DataBuffer = require('./data-buffer');