+++ src/data-stream.js
@@ -1,3 +1,4 @@
+/** @type {Function} */
 let debug = () => {}; /* istanbul ignore next */ if (process.env.UTTORI_DATA_DEBUG) { try { debug = require('debug')('DataStream'); } catch {} }
 const DataBuffer = require('./data-buffer');
 const DataBufferList = require('./data-buffer-list');